requirements.txt
streamlit
Pillow
watchdog
requests
python-dotenv
